# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Queue & Appointment API for Algerian Public Services
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /offices:
    get:
      operationId: listOffices
      tags: [services]
      summary: List offices with optional filters
      parameters:
        - name: wilayaId
          in: query
          schema:
            type: integer
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of offices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Office"
  /offices/{officeId}:
    get:
      operationId: getOffice
      tags: [services]
      summary: Get office details
      parameters:
        - name: officeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Office details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OfficeDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /offices/{officeId}/services:
    get:
      operationId: listOfficeServices
      tags: [services]
      summary: List services offered by an office
      parameters:
        - name: officeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Office services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OfficeService"
  /offices/{officeId}/available-slots:
    get:
      operationId: getAvailableSlots
      tags: [appointments]
      summary: Get available appointment slots for an office
      parameters:
        - name: officeId
          in: path
          required: true
          schema:
            type: integer
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: serviceId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Available time slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeSlot"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]
...
    Office:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
        categoryId:
          type: integer
        categoryName:
          type: string
        wilayaId:
          type: integer
        wilayaName:
          type: string
        address:
          type: string
        phone:
          type: string
        openTime:
          type: string
        closeTime:
          type: string
        workingDays:
          type: array
          items:
            type: string
        averageWaitMinutes:
          type: integer
        todayQueueCount:
          type: integer
        rating:
          type: number
      required: [id, name, nameAr, categoryId, categoryName, wilayaId, wilayaName, address, phone, openTime, closeTime, workingDays, averageWaitMinutes, todayQueueCount, rating]
    OfficeDetail:
      allOf:
        - $ref: "#/components/schemas/Office"
        - type: object
          properties:
            commune:
              type: string
            description:
              type: string
            services:
              type: array
              items:
                $ref: "#/components/schemas/OfficeService"
          required: [commune, description, services]
    OfficeService:
      type: object
      properties:
        id:
          type: integer
        officeId:
          type: integer
        nameAr:
          type: string
        nameFr:
          type: string
        duration:
          type: integer
          description: Duration in minutes
        requiredDocuments:
          type: array
          items:
            type: string
        fee:
          type: number
      required: [id, officeId, nameAr, nameFr, duration, requiredDocuments, fee]
    TimeSlot:
      type: object
      properties:
        time:
          type: string
        available:
          type: boolean
        remainingCapacity:
          type: integer
      required: [time, available, remainingCapacity]
...