# lib/api-zod/src/generated/types/office.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Queue & Appointment API for Algerian Public Services
 * OpenAPI spec version: 0.1.0
 */

export interface Office {
  id: number;
  name: string;
  nameAr: string;
  categoryId: number;
  categoryName: string;
  wilayaId: number;
  wilayaName: string;
  address: string;
  phone: string;
  openTime: string;
  closeTime: string;
  workingDays: string[];
  averageWaitMinutes: number;
  todayQueueCount: number;
  rating: number;
}