# lib/api-zod/src/generated/types/officeDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Queue & Appointment API for Algerian Public Services
 * OpenAPI spec version: 0.1.0
 */
import type { Office } from "./office";
import type { OfficeService } from "./officeService";

export type OfficeDetail = Office & {
  commune: string;
  description: string;
  services: OfficeService[];
};