# lib/api-zod/src/generated/types/officeService.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Queue & Appointment API for Algerian Public Services
 * OpenAPI spec version: 0.1.0
 */

export interface OfficeService {
  id: number;
  officeId: number;
  nameAr: string;
  nameFr: string;
  /** Duration in minutes */
  duration: number;
  requiredDocuments: string[];
  fee: number;
}